pyproject.toml
[tool.poetry]
name = "discord_university"
version = "0.1.0"
description = ""
authors = ["Joshua Salzedo <joshuasalzedo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
toml = "^0.10.2"
attrs = "^20.3.0"
cattrs = "^1.1.2"
click = "^7.1.2"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.19.0"
black = "^20.8b1"
"discord.py-stubs" = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
